+++ ddtelemetry-ffi/Cargo.toml
@@ -11,6 +11,7 @@ crate-type = ["lib", "staticlib", "cdylib"]
 
 [dependencies]
 ddtelemetry = { path = "../ddtelemetry" }
+datadog-ipc = { path = "../ipc" }
 ddcommon-ffi = { path = "../ddcommon-ffi" }
 paste = "1"
 libc = "0.2"
